update with Josh's changes
