+++ net/uring/file_test.go
@@ -1,6 +1,7 @@
 package uring
 
 import (
+	"io/ioutil"
 	"os"
 	"testing"
 
@@ -30,4 +31,25 @@ func TestFileRead(t *testing.T) {
 	n, err := uf.Read(buf)
 	c.Assert(err, qt.IsNil)
 	c.Assert(buf[:n], qt.DeepEquals, want)
+}
+
+func TestFileWrite(t *testing.T) {
+	if !Available() {
+		t.Skip("io_uring not available")
+	}
+	c := qt.New(t)
+	tmpFile, err := ioutil.TempFile(".", "uring-test")
+	c.Assert(err, qt.IsNil)
+	t.Cleanup(func() {
+		os.Remove(tmpFile.Name())
+	})
+	f, err := newFile(tmpFile)
+	c.Assert(err, qt.IsNil)
+	content := []byte("a test string to check writing works 😀 with non-unicode input")
+	n, err := f.Write(content)
+	if n != len(content) {
+		t.Errorf("mismatch between written len and content len: want %d, got %d", len(content), n)
+	}
+	c.Assert(err, qt.IsNil)
+	c.Assert(f.Close(), qt.IsNil)
 }
